+++ docs/whats-new.rst
@@ -67,6 +67,10 @@ Enhancements
   ``graphics.plot_modeloutput_map(gdirs, extend_plot_limits=True)``
   (:pull:`1508`).
   By `Patrick Schmitt <https://github.com/pat-schmitt>`_
+- Added new argument ``add_fixed_geometry_spinup`` to extend the model run of
+  ``run_dynamic_spinup`` with a fixed-geometry-spinup if the spinup period is
+  shortened(:pull:`1514`)
+  By `Patrick Schmitt <https://github.com/pat-schmitt>`_
 
 Bug fixes
 ~~~~~~~~~